% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubeRoot.R
\name{CubeRoot}
\alias{CubeRoot}
\title{Compute the non negative cube root of a real number}
\usage{
CubeRoot(x)
}
\arguments{
\item{x}{a non negative number}
}
\description{
Compute the non negative cube root of a real number
}
