% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Der.R
\name{Der}
\alias{Der}
\title{Function that approximates the derivative of a function as a weighted average of the left and right derivatives}
\usage{
Der(x, y)
}
\arguments{
\item{x}{vector of strictly increasing real numbers}

\item{y}{vector of strictly increasing real numbers corresponding to f(x)}
}
\description{
Function that approximates the derivative of a function as a weighted average of the left and right derivatives
}
