% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctMean.R
\name{FctMean}
\alias{FctMean}
\title{Function that approximates the mean of a function using the mean of the left and right Riemann sums}
\usage{
FctMean(x, y)
}
\arguments{
\item{x}{vector of strictly increasing real numbers}

\item{y}{vector of strictly increasing real numbers corresponding to f(x)}
}
\description{
Function that approximates the mean of a function using the mean of the left and right Riemann sums
}
