% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First.R
\name{First}
\alias{First}
\title{Function that returns the first non-NA value of a vector}
\usage{
First(v)
}
\arguments{
\item{v}{vector of real numbers}
}
\description{
Function that returns the first non-NA value of a vector
}
