% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstMode.R
\name{FirstMode}
\alias{FirstMode}
\title{Function that returns the mode of a vector x. In case of a tie, it returns the first value of x that is a mode}
\usage{
FirstMode(x)
}
\arguments{
\item{x}{vector of real numbers}
}
\description{
Function that returns the mode of a vector x. In case of a tie, it returns the first value of x that is a mode
}
