% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_adjacency.R
\name{knn_adjacency}
\alias{knn_adjacency}
\title{Construct a similarity matrix}
\usage{
knn_adjacency(X, K)
}
\arguments{
\item{X}{matrix X of "proximities" between the trajectories. These are non-negative numbers with the interpretation that a small proximity between two trajectories means a large similarity.}

\item{K}{The number of nearest neighbors to consider for each trajectories}
}
\description{
Define a function that construct a sparse similarity matrix (aka adjacency matrix) based on the K-nearest neighbor (KNN) principle. It takes as input a matrix X of "proximities" between points as well as the number K of neighbors to consider and outputs a sparse matrix where cell (i,j) is 0 if neither point i nor point j is among the KNN of the other, 1 if i is among the KNN of j and vice versa, and 1/2 otherwise (i.e. if i is among the KNN of j \emph{or} vice versa, but not both)
}
