% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trajClusters.R
\name{plot.trajClusters}
\alias{plot.trajClusters}
\alias{trajScatter}
\alias{CVIplot}
\title{Plot \code{trajClusters} object}
\usage{
\method{plot}{trajClusters}(x, sample.size = 5, ask = TRUE, which.plots = NULL, ...)

trajScatter(x, ask = TRUE, which.scatter = NULL, N = NULL, ...)

CVIplot(x, ...)
}
\arguments{
\item{x}{object of class \code{trajClusters} as returned by the function
\code{trajClusters()}.}

\item{sample.size}{the number of trajectories to be randomly sampled
from each clusters. If \code{NULL}, all the trajectories are included in a random order. Defaults to \code{5}.}

\item{ask}{logical. If \code{TRUE}, the user is asked before each plot. Defaults to
\code{TRUE}.}

\item{which.plots}{either \code{NULL} or a vector of integers. If \code{NULL}, every
available plot is displayed. If a vector is supplied, only the corresponding
plots will be displayed.}

\item{...}{other parameters to be passed through to plotting functions.}

\item{which.scatter}{either \code{NULL} or a vector of integers that is a subset of the \code{measure} argument used in the function \code{trajClusters} to produce object \code{x}. If \code{NULL}, every available scatter plots are displayed. If a vector is supplied, only the corresponding plots will be displayed.}

\item{N}{the maximum number of points present in each scatter plots. If a non \code{NULL} value is specified, \code{N} points are sampled randomly in a way that preserves the relative groups sizes. If \code{NULL} (the default), all the points are plotted.}
}
\description{
\code{plot()} sequentially plots (i) the clusters centroids, (ii) a random sample of trajectories from each cluster, and (iii) the median of each standardized measure by cluster. \code{(trajScatter()} displays the two dimensional projections of the standardized data in feature space for every pair of measures. \code{CVIplot()} graphs the internal cluster validity indices as a function of the number of clusters as well as the result of the ranked voting system for determining the optimal number of clusters.
}
\examples{
data(trajdata)

m = trajMeasures(trajdata[, -2], ID = TRUE)

c3 = trajClusters(m, nclusters = 3)

plot(c3, which.plots = 1, ask = FALSE) ## centroids

plot(c3, which.plots = 2, ask = FALSE) ## sample trajectories

plot(c3, which.plots = 3, ask = FALSE) ## median of standardized measures by cluster



}
