% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{Function to prevent another function from printing things as part of its execution}
\usage{
quiet(x)
}
\arguments{
\item{x}{code}
}
\description{
Function to prevent another function from printing things as part of its execution
}
