% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{spect}
\alias{spect}
\title{Implements spectral clustering}
\usage{
spect(x, k, nstart, fuzzy)
}
\arguments{
\item{x}{trajectories to be clustered}

\item{k}{number of clusters}

\item{nstart}{how many random starts to use in the underlying k-means algorithm}

\item{fuzzy}{logical; TRUE if fuzzy k-means is to be used instead of ordinary k-means, so as to yield a fuzzy partition.}
}
\description{
Create a sparse similarity matrix and applies to it the spectral clustering algorithm described in Meila (2005)
}
