% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-package.R
\docType{package}
\name{traj-package}
\alias{traj}
\alias{traj-package}
\title{traj: Feature-Based Clustering of Longitudinal Trajectories}
\description{
Identifies clusters of individual longitudinal trajectories. In the spirit of Leffondre et al. (2004), the procedure involves identifying each trajectory to a point in the space of measures. In this context, a measure is a quantity meant to capture a certain characteristic feature of the trajectory. The points in the space of measures are then clustered using a version of the Spectral Clustering algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=traj}
}

}
\author{
\strong{Maintainer}: Laurence Boulanger \email{laurence.boulanger@umontreal.ca}

Authors:
\itemize{
  \item Laurence Boulanger \email{laurence.boulanger@umontreal.ca}
  \item Marie-Pierre Sylvestre \email{marie-pierre.sylvestre@umontreal.ca}
}

Other contributors:
\itemize{
  \item Jean-Benoit Bergeron [contributor]
  \item Gillis Delmas [contributor]
  \item Tchouangue Dinkou [contributor]
  \item Dan Vatnik [contributor]
}

}
\keyword{internal}
