% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajClusters-methods.R
\name{trajAnalysis}
\alias{trajAnalysis}
\title{Perform order analysis on the cluster-wise medians}
\usage{
trajAnalysis(object, top_p = 3)
}
\arguments{
\item{object}{An object of class \code{trajClusters}.}

\item{top_p}{The \code{top_p} most discriminating measures for each cluster to be reported in the summary.}
}
\description{
Identifies which measures are most extreme for each cluster of a \code{trajClusters} object.
}
\details{
For each cluster, the median of each standardized measure is computed and the difference (\code{delta}) between this median and the median of the medians is computed. A large positive or negative value of this difference is interpreted to mean that the measure is ``extreme'' for the cluster. The \code{top_p} most extreme measures for each cluster are reported under the column \code{measure}. The \code{delta} column logs the value of the difference. The number under the column \code{rank} is 1 if the absolute value of \code{delta} is the largest among all clusters, it is 2 if the absolute value of \code{delta} is the second largest among all clusters, etc. The \code{direction} column reads \code{largest} (resp. \code{smallest}) if the cluster's median is largest (resp. smallest) among all clusters. Otherwise, it reads \code{large} (resp. \code{small}) if \code{delta} is positive (resp. negative).
}
