% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajClusters.R
\name{trajClusters}
\alias{trajClusters}
\alias{print.trajClusters}
\alias{summary.trajClusters}
\title{Classify the Longitudinal Data Based on the Measures}
\usage{
trajClusters(
  Measures,
  select = NULL,
  fuzzy = FALSE,
  nclusters = NULL,
  subset.n = NULL,
  nstart = 50
)

\method{print}{trajClusters}(x, ...)

\method{summary}{trajClusters}(object, ...)
}
\arguments{
\item{Measures}{object of class \code{trajMeasures} as returned by the function
\code{trajMeasures()}.}

\item{select}{an optional vector of positive integers corresponding to the
measures to use in the clustering. Defaults to \code{NULL}, which uses all the measures contained in \code{Measures}.}

\item{fuzzy}{logical. If FALSE, each trajectory is assigned to a unique group. If TRUE, each trajectory is assigned a "degree of membership" to each group. Defaults to FALSE.}

\item{nclusters}{The desired number of clusters. If \code{NULL}, clustering is carried out for every number of clusters between 2 and (up to) 8 and the "best" number of clusters is used, as judged by the combination of three internal cluster validity indices. See section 'Value' for more details. Defaults to \code{NULL}.}

\item{subset.n}{A positive numerical integer smaller than the number of trajectories. If \code{nclusters} is \code{NULL}, \code{subset} is the number of trajectories, randomly sampled from the complete data set, that will be used to determine the optimal number of clusters in the interest of speeding up the process.}

\item{nstart}{The number of random starts. Defaults to \code{50}.}

\item{x}{object of class \code{trajClusters}.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{object of class \code{trajClusters}.}
}
\value{
An object of class \code{trajClusters}; a list containing the result of the clustering, as well as a curated form of the arguments.
}
\description{
Classifies the trajectories by applying a nonparametric clustering algorithm to the measures computed by \code{trajMeasures()}.
}
\details{
The spectral clustering algorithm presented in Meila (2005) is implemented in which the similarity matrix \eqn{S} is built from a binary K nearest neighbors similarity function (\eqn{S=(W+W^T)/2}, where \eqn{W_{ij}=1} if data point \eqn{j} is among the nearest points to data point \eqn{i} and \eqn{W_{ij}=0} otherwise).

When \code{nclusters = NULL}, the function evaluates candidate clusterings with number of clusters \eqn{k} ranging from 2 to 8 using three internal validity indices: C-index, Calinski-Harabasz and Wemmert-Gançarski. These indices are normalized so that the highest value is 1 and the lowest is 0, and so that a high value is synonymous with high validity. The optimal number of clusters is determined according to a ranked voting system in which each index contributes a fractional vote according to its ranking of the candidate solutions. Specifically, each index casts a vote worth 1 in favor of \eqn{k} if it takes its greatest value when the number of groups is \eqn{k}, worth 5/6 if it takes its second greatest value when the number of groups is \eqn{k}, and so on down to a vote worth 0 if the index takes its smallest value when the number of groups is \eqn{k}. The favorability of \eqn{k} is the sum of the 3 votes.
}
\examples{

data(trajdata)

dat <- trajdata[, -2] #remove the Group column

m <- trajMeasures(dat, ID = TRUE, measures = c(1:20))

s <- trajClusters(m, nclusters = 4)
plot(s, which.plots = 2, ask = FALSE)

dat$clusters <- s$partition[match(trajdata$id, s$partition[, 1]), 2]
tail(dat)


}
\references{
Meila, M., Spectral Clustering. Handbook of Cluster Analysis, Chapter 7, Chapman and Hall/CRC, 2005.
}
