% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajClusters-methods.R
\name{trajClusterSummary}
\alias{trajClusterSummary}
\title{Computes the cluster-wise summaries of the measures}
\usage{
trajClusterSummary(object)
}
\arguments{
\item{object}{An object of class \code{trajClusters}.}
}
\value{
A list where each entry contains the summaries of the measures for the trajectories in the corresponding cluster
}
\description{
Computes the cluster-wise summaries of the measures from a
\code{trajClusters} object.
}
\examples{
data(trajdata)
m <- trajMeasures(trajdata[, -2], ID = TRUE, measures = 1:20)
c <- trajClusters(m, nclusters = 4, fuzzy = TRUE)
trajClusterSummary(c)

}
