% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trajdata}
\alias{trajdata}
\title{Example Longitudinal Trajectory Data}
\format{
This data frame has 130 rows and the following 8 columns:
\describe{
\item{id}{An identification variable that runs from 1 to 130.}
\item{Group}{A character variable that's either "A", "B", "C" or "D" depending on which of the four data generating process the trajectory is coming from.}
\item{t1}{The observation of the trajectory at time t = 1.}
\item{t2}{The observation of the trajectory at time t = 2.}
\item{t3}{The observation of the trajectory at time t = 3.}
\item{t4}{The observation of the trajectory at time t = 4.}
\item{t5}{The observation of the trajectory at time t = 5.}
\item{t6}{The observation of the trajectory at time t = 6.}
}
}
\usage{
trajdata
}
\description{
An artificially created data set with 130 trajectories split into four groups, labelled A, B, C, D according to the data generating process.
}
\keyword{datasets}
