% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajClusters-methods.R
\name{trajFuzzyPartition}
\alias{trajFuzzyPartition}
\title{Extract Fuzzy Cluster Memberships}
\usage{
trajFuzzyPartition(object)
}
\arguments{
\item{object}{An object of class \code{trajClusters}.}
}
\value{
A matrix containing the membership value of each trajectory
in each cluster. Each row corresponds to a trajectory and each
column to a cluster. The values in each row represent the degree
of membership of the trajectory in the corresponding clusters.
Returns \code{NULL} if the clustering was performed with
\code{fuzzy = FALSE}.
}
\description{
Extracts the fuzzy cluster membership matrix from a
\code{trajClusters} object.
}
\examples{
data(trajdata)
m <- trajMeasures(trajdata[, -2], ID = TRUE, measures = 1:20)
c <- trajClusters(m, nclusters = 4, fuzzy = TRUE)
trajFuzzyPartition(c)

}
