% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajReduce-methods.R
\name{trajFuzzyReducedPartition}
\alias{trajFuzzyReducedPartition}
\title{Extract Fuzzy Cluster Memberships}
\usage{
trajFuzzyReducedPartition(object)
}
\arguments{
\item{object}{An object of class \code{trajReduce}.}
}
\description{
Extracts the fuzzy cluster membership matrix from a
\code{trajReduce} object.
}
