% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajMouseover.R
\name{trajMouseover}
\alias{trajMouseover}
\title{Interactive Plot of Trajectories by Cluster}
\usage{
trajMouseover(x)
}
\arguments{
\item{x}{object of class \code{trajClusters} as returned by the function
\code{trajClusters()}.}
}
\value{
An interactive \code{highchart} object.
}
\description{
Creates an interactive \code{highcharter} plot displaying
trajectories colored according to their cluster membership.
}
\details{
The plot provides interactive highlighting: when the user hovers over a
trajectory, only that trajectory is highlighted while the other trajectories
become less prominent. Clicking
on a cluster in the legend hides or displays all trajectories belonging to
that cluster. Zoom vertically by clicking and dragging.
}
\examples{
data(trajdata)

m = trajMeasures(trajdata[, -2], ID = TRUE)

c3 = trajClusters(m, nclusters = 3)

trajMouseover(c3)

}
