% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajClusters-methods.R
\name{trajPartition}
\alias{trajPartition}
\title{Extract Cluster Assignments}
\usage{
trajPartition(object)
}
\arguments{
\item{object}{An object of class \code{trajClusters}.}
}
\value{
A data frame with two columns: \code{ID}, containing the
trajectory identifiers, and \code{Cluster}, containing the cluster
assignment of each trajectory.
}
\description{
Extracts the cluster assignment of each trajectory from a
\code{trajClusters} object.
}
\examples{
data(trajdata)
m <- trajMeasures(trajdata[, -2], ID = TRUE, measures = 1:20)
c <- trajClusters(m, nclusters = 4)
trajPartition(c)

}
