% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajReduce.R
\name{trajReduce}
\alias{trajReduce}
\alias{print.trajReduce}
\title{Select a Subset of the Measures Using a Similarity Criterion on the Set of Clusterings}
\usage{
trajReduce(Measures, Clusters, crit = "ARI", keep = 3)

\method{print}{trajReduce}(x, ...)
}
\arguments{
\item{Measures}{object of class \code{trajMeasures} as returned by \code{\link[traj]{trajMeasures}}.}

\item{Clusters}{object of class \code{trajClusters} as returned by \code{\link[traj]{trajClusters}}.}

\item{crit}{The similarity criterion Either "ARI" for the Adjusted Rand Index of Hubert and Arabie (1985), "nVId" for the normalized variation of information distance (eg. Meila (2007)) or "nSJd" for the normalized split/joint distance of van Dongen (2000).}

\item{keep}{The number of measures to keep. Defaults to 3.}

\item{x}{object of class \code{trajReduce}.}
}
\value{
An object of class \code{trajReduce}; a list containing the reduced list of measures, the corresponding reduced clustering (and fuzzy clustering if applicable) as well as the value of the similarity between the reduced clustering and the initial clustering.
}
\description{
This function examines the effect of reducing the number of measures on which the trajectories are clustered. Specifically, starting from a clustering \eqn{C} in the form of an object of class \code{trajClusters} and a choice of a similarity criterion to compare clusterings, this function finds the subset of measures which results in the clustering most similar to \eqn{C}.
}
\details{
The Rand index ranges from 0 to 1 with 0 indicating identical clusters and 1 indicating maximally different clusters. The normalized variation of information distance (nVId) and normalized split-join distance (nSJd) and have the opposite interpretation with 0 indicating maximally different clusters and 1 indicating identical clusters. Therefor, to facilitate comparison, we plot 1 - nVId (resp. 1 - nSJd) instead of nVId (resp. nSJd).

The function runs the \code{\link[traj]{trajClusters}} function on every combination of \code{keep} measures and identifies the reduced representation yielding the clustering most similar to the (hard) clustering in \code{Clusters}. If the (hard) clustering in \code{Clusters} derives from a soft clustering, the iterations of \code{\link[traj]{trajClusters}}  will also be ran using \code{fuzzy = TRUE} and the corresponding hard clusterings will be used for comparison.
}
\examples{
\donttest{
data(trajdata)

dat <- trajdata[, -2] #remove the Group column

m <- trajMeasures(dat, ID = TRUE, measures = c(1:20))

s <- trajClusters(m, nclusters = 4)

r <- trajReduce(m, s, keep = 2)
}

}
\references{
Hubert L, Arabie P. Comparing partitions. Journal of Classification 2:193-218, 1985.

Meila M. Comparing clusterings -- an information based distance. Journal of Multivariate Analysis, 98, pp 873-895, 2007.

van Dongen S. Performance criteria for graph clustering and Markov cluster experiments. Technical Report INS-R0012, National Research Institute for Mathematics and Computer Science in the Netherlands, Amsterdam, May 2000.
}
