% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Last.R
\name{Last}
\alias{Last}
\title{Function that returns the last non-NA value of a vector}
\usage{
Last(v)
}
\arguments{
\item{x}{vector of real numbers}
}
\description{
Function that returns the last non-NA value of a vector
}
