% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trajClusters.R
\name{plot.trajClusters}
\alias{plot.trajClusters}
\alias{scatterplots}
\alias{CVIplot}
\title{Plot \code{trajClusters} object}
\usage{
\method{plot}{trajClusters}(x, sample.size = 5, ask = TRUE, which.plots = NULL, ...)

scatterplots(x, ask = TRUE, which.scatter = NULL, N = NULL, ...)

CVIplot(x, ...)
}
\arguments{
\item{x}{object of class \code{trajClusters} as returned by the function
\code{trajClusters()}.}

\item{sample.size}{the number of trajectories to be randomly sampled
from each clusters. If \code{NULL}, all the trajectories are included in a random order. Defaults to \code{5}.}

\item{ask}{logical. If \code{TRUE}, the user is asked before each plot. Defaults to
\code{TRUE}.}

\item{which.plots}{either \code{NULL} or a vector of integers. If \code{NULL}, every
available plot is displayed. If a vector is supplied, only the corresponding
plots will be displayed.}

\item{...}{other parameters to be passed through to plotting functions.}

\item{which.scatter}{either \code{NULL} or a vector of integers that is a subset of the \code{measure} argument used in function \code{trajClusters} to produce object \code{x}. If \code{NULL}, every available scatter plots are displayed. If a vector is supplied, only the corresponding plots will be displayed.}

\item{N}{the maximum number of points present in each scatter plots. If a non \code{NULL} value is specified, N points are sampled randomly in a way that preserves the relative groups sizes. If \code{NULL} (the default), all the points are plotted.}
}
\description{
Plots the curves corresponding to (or closest to) the centroids of the clusters and plots a random sample from each groups.
}
\examples{
\dontrun{
data("trajdata")
trajdata.noGrp <- trajdata[, -which(colnames(trajdata) == "Group")] #remove the Group column

m = trajMeasures(trajdata.noGrp, ID = TRUE)
c3 = trajClusters(m, nclusters = 3)

plot(c3)

#The pointwise mean trajectories correspond to the third and fourth displayed plots.

c4 = trajClusters(m, nclusters = 4)

plot(c4, which.plots = 3:4)

}


}
