% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-package.R
\docType{package}
\name{traj-package}
\alias{traj}
\alias{traj-package}
\title{traj: Trajectory Analysis}
\description{
Implements the three-step procedure proposed by Leffondree et al. (2004) to identify clusters of individual longitudinal trajectories. The procedure involves (1) calculating 24 measures describing the features of the trajectories; (2) using factor analysis to select a subset of the 24 measures and (3) using cluster analysis to identify clusters of trajectories, and classify each individual trajectory in one of the clusters.
}
\author{
\strong{Maintainer}: Marie-Pierre Sylvestre \email{marie-pierre.sylvestre@umontreal.ca}

Authors:
\itemize{
  \item Marie-Pierre Sylvestre
  \item Dan Vatnik
}

}
\keyword{internal}
