\name{trajdata}
\alias{trajdata}
\docType{data}
\title{
Fictitious functional data for illustration purposes in \code{traj}.
}
\description{
Fictitious functional data for illustration purposesin \code{traj}.
}
\usage{data("trajdata")}
\format{
  This data set contains 130 rows and 7 columns, the first of which is an ID variable. The other columns represent six measurements \eqn{f(x_1),\ldots,f(x_6)} of some function \eqn{f} made at \eqn{x_1<\ldots<x_6}.
}
\keyword{datasets}
